+++ 4Diy/reduction/reduction_gpu.cuh
@@ -1,5 +1,5 @@
 
 
-void warnup_gpu( unsigned int* array, int size );
+void warnup_gpu( float* array, int size );
 
-unsigned int reduction_gpu( unsigned int* array, int size );
+float reduction_gpu( float* array, int size );
